requirements.txt
pandas
dnspython
pyspf
tqdm